add lsp rename popup window
thx to @LeonHeidelbach for fixing some outdated stuff in the original table : https://www.reddit.com/r/neovim/comments/ql4iuj/rename_hover_including_window_title_and/

+++ lua/core/mappings.lua
@@ -147,7 +147,7 @@ M.lspconfig = {
 
       ["<leader>ra"] = {
          function()
-            vim.lsp.buf.rename()
+            vim.lsp.buf.rename.float()
          end,
          "   lsp rename",
       },
